Maintenance
Just adding some comments, also splitting responsibilities
