+++ dataHandlers/generateSample.js
@@ -1,3 +1,7 @@
 import { generateSample } from './fileHandler.js';
 
+/*
+  Just an intermediary handler to generate samples
+*/
+
 generateSample();
